fix: избавление от циклов

+++ ewm-service/src/main/java/ru/practicum/explore/events/dto/EventFullDto.java
@@ -21,7 +21,7 @@ public class EventFullDto {
     private Integer id;
     private String annotation;
     private CategoryDtoResponse category;
-    private Integer confirmedRequests;
+    private Long confirmedRequests;
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss")
     private LocalDateTime createdOn;
     @NotBlank
@@ -38,5 +38,5 @@ public class EventFullDto {
     private State state;
     private String title;
     private Long views;
-    private Integer comments;
+    private Long comments;
 }
